create cities
