+++ src/components/cities/cities.jsx
@@ -1,0 +1,58 @@
+import React from "react";
+import PropTypes from "prop-types";
+import PlaceCard from "../place-card/place-card.jsx";
+
+const Cities = (props) => {
+
+  const {available, descriptions, onHeadlineButtonClick} = props;
+
+  const places = descriptions.map((description) =>
+    <PlaceCard
+      key={description}
+      description={description}
+      onHeadlineButtonClick={onHeadlineButtonClick}
+    />);
+
+  return (
+    <div className="cities">
+      <div className="cities__places-container container">
+        <section className="cities__places places">
+          <h2 className="visually-hidden">Places</h2>
+          <b className="places__found">{available} places to stay in Amsterdam</b>
+          <form className="places__sorting" action="#" method="get">
+            <span className="places__sorting-caption">Sort by</span>
+            <span className="places__sorting-type" tabIndex="0">
+              Popular
+              <svg className="places__sorting-arrow" width="7" height="4">
+                <use xlinkHref="#icon-arrow-select"></use>
+              </svg>
+            </span>
+            <ul className="places__options places__options--custom places__options--opened">
+              <li className="places__option places__option--active" tabIndex="0">Popular</li>
+              <li className="places__option" tabIndex="0">Price: low to high</li>
+              <li className="places__option" tabIndex="0">Price: high to low</li>
+              <li className="places__option" tabIndex="0">Top rated first</li>
+            </ul>
+          </form>
+          <div className="cities__places-list places__list tabs__content">
+            {places}
+          </div>
+        </section>
+        <div className="cities__right-section">
+          <section className="cities__map map"></section>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+Cities.propTypes = {
+  available: PropTypes.number.isRequired,
+  descriptions: PropTypes.arrayOf(
+      PropTypes.string)
+  .isRequired,
+  onHeadlineButtonClick: PropTypes.func.isRequired,
+};
+
+export default Cities;
+
