+++ src/components/main/main.jsx
@@ -1,18 +1,12 @@
 import React from "react";
 import PropTypes from "prop-types";
-import PlaceCard from "../place-card/place-card.jsx";
 import Header from "../header/header.jsx";
 import Locations from "../locations/locations.jsx";
+import Cities from "../cities/cities.jsx";
 
 const Main = (props) => {
-  const {available, descriptions, onHeadlineButtonClick} = props;
 
-  const places = descriptions.map((description) =>
-    <PlaceCard
-      key={description}
-      description={description}
-      onHeadlineButtonClick={onHeadlineButtonClick}
-    />);
+  const {available, descriptions, onHeadlineButtonClick} = props;
 
   return (
     <div className="page page--gray page--main">
@@ -26,35 +20,13 @@ const Main = (props) => {
           {<Locations />}
 
         </div>
-        <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{available} places to stay in Amsterdam</b>
-              <form className="places__sorting" action="#" method="get">
-                <span className="places__sorting-caption">Sort by</span>
-                <span className="places__sorting-type" tabIndex="0">
-                  Popular
-                  <svg className="places__sorting-arrow" width="7" height="4">
-                    <use xlinkHref="#icon-arrow-select"></use>
-                  </svg>
-                </span>
-                <ul className="places__options places__options--custom places__options--opened">
-                  <li className="places__option places__option--active" tabIndex="0">Popular</li>
-                  <li className="places__option" tabIndex="0">Price: low to high</li>
-                  <li className="places__option" tabIndex="0">Price: high to low</li>
-                  <li className="places__option" tabIndex="0">Top rated first</li>
-                </ul>
-              </form>
-              <div className="cities__places-list places__list tabs__content">
-                {places}
-              </div>
-            </section>
-            <div className="cities__right-section">
-              <section className="cities__map map"></section>
-            </div>
-          </div>
-        </div>
+
+        {<Cities
+          available={available}
+          descriptions={descriptions}
+          onHeadlineButtonClick={onHeadlineButtonClick}
+        />}
+
       </main>
     </div>
   );
